+++ Someren-master/SomerenModel/Drink.cs
@@ -11,7 +11,7 @@ namespace SomerenModel
         public int ProductID { get; set; } // auto incremented in SQL
         public string DrinkName { get; set; }
         public int StockAmount { get; set; } // how many in stock
-        public double SalesValue { get; set; } // price of total drinks
+        public int SalesValue { get; set; } // price of total drinks
         public int NumberOfDrinksSold { get; set; } // how many of the drinks are sold for each product
         
     }
